Starting vim works

+++ kernel/include/sys_ioctl.h
@@ -1,3 +1,3 @@
 #pragma once
 #include "klib/stdint.h"
-int sys_ioctl(int fd, uint32_t request, void *arg);
+int sys_ioctl(int fd, unsigned long request, void *arg);
